feat/add dinamic import for stats chart to improve performance
