+++ src/components/features/TypeBox/Stats.js
@@ -1,17 +1,4 @@
-import React, { useEffect, useState } from "react";
-import { Tooltip } from "@mui/material";
-import { CHAR_TOOLTIP_TITLE } from "../../../constants/Constants";
-import {
-  Line,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip as TooltipChart,
-  ResponsiveContainer,
-  Bar,
-  ComposedChart,
-} from "recharts";
-import { red } from "@mui/material/colors";
+import React, { Suspense, useEffect, useState } from "react";
 
 const Stats = ({
   status,
@@ -27,6 +14,7 @@ const Stats = ({
   incorrectCharsCount,
 }) => {
   const [roundedRawWpm, setRoundedRawWpm] = useState(0);
+  const [dinamiStatsChart, setDinamicStatsChart] = useState(null);
   const roundedWpm = Math.round(wpm);
 
   useEffect(() => {
@@ -66,9 +54,29 @@ const Stats = ({
     error: history.error,
   }));
 
+  const loadDinamicStatsChart = async () => {
+    const module = await import("../../common/StatsChart");
+    const StatsChart = module.default;
+    setDinamicStatsChart(
+      <StatsChart
+        accuracy={accuracy}
+        countDownConstant={countDownConstant}
+        language={language}
+        data={data}
+        theme={theme}
+        rawKeyStrokes={rawKeyStrokes}
+        renderResetButton={renderResetButton}
+        statsCharCount={statsCharCount}
+      />
+    );
+  };
+
   useEffect(() => {
     if (status === "started") {
       setTypingTestHistory(initialTypingTestHistory);
+    }
+    if (status === "finished") {
+      loadDinamicStatsChart();
     }
   }, [status]);
 
@@ -104,169 +112,6 @@ const Stats = ({
     }
   }, [countDown]);
 
-  const getFormattedLanguageLanguageName = (value) => {
-    switch (value) {
-      case "ENGLISH_MODE":
-        return "eng";
-      case "CHINESE_MODE":
-        return "chn";
-      default:
-        return "eng";
-    }
-  };
-
-  const renderCharStats = () => (
-    <Tooltip
-      title={
-        <span style={{ whiteSpace: "pre-line" }}>{CHAR_TOOLTIP_TITLE}</span>
-      }
-    >
-      <div>
-        <p className="stats-title">Characters</p>
-        <h2 className="stats-value">
-          <span className="correct-char-stats">{statsCharCount[1]}</span>/
-          <span className="incorrect-char-stats">{statsCharCount[2]}</span>/
-          <span className="missing-char-stats">{statsCharCount[3]}</span>/
-          <span className="correct-char-stats">{statsCharCount[4]}</span>/
-          <span className="incorrect-char-stats">{statsCharCount[5]}</span>
-        </h2>
-      </div>
-    </Tooltip>
-  );
-
-  const renderIndicator = (color) => (
-    <span
-      style={{ backgroundColor: color, height: "12px", width: "24px" }}
-    ></span>
-  );
-
-  const CustomTooltip = ({ active, payload, label }) => {
-    if (active && payload && payload.length) {
-      const payloadData = payload[0].payload;
-      return (
-        <div
-          className="custom-tooltip"
-          style={{
-            paddingInline: "8px",
-            paddingBlock: "2px",
-          }}
-        >
-          <p className="label" style={{ fontSize: "12px", fontWeight: "bold" }}>
-            {`Time: ${label} s`}
-          </p>
-          <p className="desc tooltip">
-            {renderIndicator(red[400])}
-            {`Errors: ${payloadData.error}`}
-          </p>
-          <p className="desc tooltip">
-            {renderIndicator(theme.textTypeBox)}
-            {`Raw WPM: ${payloadData.rawWpm}`}
-          </p>
-          <p className="desc tooltip">
-            {renderIndicator(theme.text)}
-            {`WPM: ${payloadData.wpm}`}
-          </p>
-        </div>
-      );
-    }
-
-    return null;
-  };
-
-  const renderAccuracy = () => (
-    <div style={{ marginTop: "16px" }}>
-      <h2 className="primary-stats-title">ACC</h2>
-      <h1 className="primary-stats-value">{accuracy}%</h1>
-    </div>
-  );
-
-  const renderRawKpm = () => (
-    <div>
-      <p className="stats-title">KPM</p>
-      <h2 className="stats-value">
-        {Math.round((rawKeyStrokes / Math.max(countDownConstant, 1)) * 60.0)}
-      </h2>
-    </div>
-  );
-
-  const renderLanguage = () => (
-    <div>
-      <p className="stats-title">Test Mode</p>
-      <h2 className="stats-value">
-        {getFormattedLanguageLanguageName(language)}
-      </h2>
-    </div>
-  );
-
-  const renderTime = () => (
-    <div>
-      <p className="stats-title">Time</p>
-      <h2 className="stats-value">{countDownConstant} s</h2>
-    </div>
-  );
-
-  const renderWpm = () => {
-    const totalWpm = data.map((e) => e.wpm).reduce((a, b) => a + b, 0);
-    const averageWpm = data.length > 1 ? totalWpm / (data.length - 1) : 0;
-    return (
-      <div>
-        <h2 className="primary-stats-title">WPM</h2>
-        <h1 className="primary-stats-value">{Math.round(averageWpm)}</h1>
-      </div>
-    );
-  };
-
-  const Chart = () => (
-    <ResponsiveContainer
-      width="100%"
-      minHeight={200}
-      maxHeight={200}
-      height="100%"
-    >
-      <ComposedChart
-        width="100%"
-        height="100%"
-        data={data.filter((d) => d.time !== 0)}
-        margin={{
-          top: 12,
-          right: 12,
-          left: 0,
-          bottom: 0,
-        }}
-      >
-        <CartesianGrid
-          vertical={false}
-          horizontal={false}
-          stroke={theme.text}
-          opacity={0.15}
-        />
-        <XAxis
-          dataKey="time"
-          stroke={theme.text}
-          tickMargin={10}
-          opacity={0.25}
-        />
-        <YAxis stroke={theme.text} tickMargin={10} opacity={0.25} />
-        <TooltipChart cursor content={<CustomTooltip />} />
-        <Line
-          type="monotone"
-          dataKey="rawWpm"
-          stroke={theme.textTypeBox}
-          dot={false}
-          activeDot={false}
-        />
-        <Line
-          type="monotone"
-          dataKey="wpm"
-          stroke={theme.text}
-          dot={false}
-          activeDot={false}
-        />
-        <Bar dataKey="error" barSize={12} fill={`${red[400]}`} />
-      </ComposedChart>
-    </ResponsiveContainer>
-  );
-
   return (
     <>
       {status !== "finished" && (
@@ -276,26 +121,7 @@ const Stats = ({
         </>
       )}
 
-      {status === "finished" && (
-        <div className="stats-overlay">
-          <section className="stats-chart">
-            <section className="stats-header">
-              <div>
-                {renderWpm()}
-                {renderAccuracy()}
-              </div>
-              {Chart()}
-            </section>
-            <section className="stats-footer">
-              {renderLanguage()}
-              {renderRawKpm()}
-              {renderCharStats()}
-              {renderTime()}
-            </section>
-            <section>{renderResetButton()}</section>
-          </section>
-        </div>
-      )}
+      {status === "finished" && <Suspense>{dinamiStatsChart}</Suspense>}
     </>
   );
 };
